+++ src/macos.rs
@@ -13,7 +13,7 @@ enum NSBitmapImageFileType {
     NSBitmapImageFileTypePNG = 4,
 }
 
-pub fn get_icon(path: &str, size: i32) -> Result<Vec<u8>, Error> {
+pub fn get_icon(path: &str, size: u16) -> Result<Vec<u8>, Error> {
     let size: f64 = size.into();
     unsafe {
         // convert &str to NSString
